requirements.txt
apache-airflow
praw
dotenv